Split driver_otp into driver and controller.
